unit test service( Account , CAtegory , Oderdetail, Product ) , check out cart UI cart, add product to cart => check out

+++ src/main/java/com/example/ecommerce_rookies/services/impl/AccountServiceImpl.java
@@ -1,5 +1,6 @@
 package com.example.ecommerce_rookies.services.impl;
 
+import com.example.ecommerce_rookies.exception.account.NotFoundAccount;
 import com.example.ecommerce_rookies.modelDTO.AccountDto;
 import com.example.ecommerce_rookies.modelDTO.SetActivity;
 import com.example.ecommerce_rookies.models.Account;
@@ -7,7 +8,9 @@ import com.example.ecommerce_rookies.models.Infomation;
 import com.example.ecommerce_rookies.repository.AccountRepository;
 import com.example.ecommerce_rookies.repository.InfomationRepository;
 import com.example.ecommerce_rookies.services.AccountService;
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -18,15 +21,17 @@ import java.util.Optional;
 
 @Service
 public class AccountServiceImpl implements AccountService {
-    @Autowired
-    private AccountRepository accountRepository;
-
-    @Autowired
-    PasswordEncoder encoder;
+    private final AccountRepository accountRepository;
 
 
-    @Autowired
-    private InfomationRepository infomationRepository;
+    private final InfomationRepository infomationRepository;
+    private final ModelMapper modelMapper;
+     @Autowired
+    public AccountServiceImpl(AccountRepository accountRepository, InfomationRepository infomationRepository, ModelMapper modelMapper) {
+        this.accountRepository = accountRepository;
+        this.infomationRepository = infomationRepository;
+        this.modelMapper = modelMapper;
+    }
 
     @Override
     public List<Account> getAccountList() {
@@ -34,11 +39,20 @@ public class AccountServiceImpl implements AccountService {
     }
 
     @Override
-    public Optional<Account> getAccountId(Long id) { return accountRepository.findById(id); }
+    public Optional<Account> getAccountId(Long id) {
+        if(accountRepository.findById(id).isEmpty())
+            throw new NotFoundAccount(id);
+        return accountRepository.findById(id);
+    }
 
     @Override
-    public void deleteAccountId(Long id){
+    public ResponseEntity<?> deleteAccountId(Long id){
+        Optional<Account> account = Optional.of(accountRepository.getReferenceById(id));
+        if(account.isEmpty())
+            throw new NotFoundAccount(id);
+        infomationRepository.deleteAllByAccount_Id(id);
         accountRepository.deleteById(id);
+        return ResponseEntity.ok().body("User delete successfully!");
     }
 
     @Override
@@ -62,35 +76,42 @@ public class AccountServiceImpl implements AccountService {
         accountDto.setEmail(infomation.get().getEmail());
         accountDto.setAvatar(infomation.get().getAvatar());
         accountDto.setPhone(infomation.get().getPhone());
+        accountDto.setRole(account.getRoles().getRolename());
         accountDto.setCheck_activity(account.isCheck_Account());
         return accountDto;
     }
 
     @Override
-    public Account updateAccount(long id, AccountDto accountDto) {
-        Account account = accountRepository.getOne(id);
-        Infomation infomation = infomationRepository.getOne(id);
-        Account user = new Account( encoder.encode(accountDto.getPassword()),accountDto.getUsername());
-        accountRepository.save(account);
+    public ResponseEntity<?> updateAccount(long id, AccountDto accountDto) {
+        Optional<Account> account = accountRepository.findById(id);
+        if(account.isEmpty())
+            throw  new NotFoundAccount(id);
+        Infomation infomation = infomationRepository.getReferenceById(id);
         infomation.setUsername(accountDto.getUsername());
         infomation.setAddress(accountDto.getAddress());
         infomation.setEmail(accountDto.getEmail());
         infomation.setAvatar(accountDto.getAvatar());
         infomation.setPhone(accountDto.getPhone());
-        infomation.setAccount(accountRepository.save(account));
+        infomation.setAccount(accountRepository.save(account.get()));
         infomationRepository.save(infomation);
-        return  accountRepository.save(account);
+        accountRepository.save(account.get());
+
+        return ResponseEntity.ok().body("User update successfully!");
     }
 
     @Override
-    public Account updateActivityAccount(long id, SetActivity setActivity) {
-        Account account = accountRepository.getOne(id);
+    public ResponseEntity<?> updateActivityAccount(long id, SetActivity setActivity) {
+        Optional<Account> account = accountRepository.findById(id);
+        if(account.isEmpty() )
+            throw  new NotFoundAccount(id);
         if(Objects.equals(setActivity.getSet_activity(), "0")) {
-            account.setCheck_Account(false);
+            account.get().setCheck_Account(false);
         }else{
-            account.setCheck_Account(true);
+            account.get().setCheck_Account(true);
         }
-        return  accountRepository.save(account);
+        accountRepository.save(account.get());
+        return ResponseEntity.ok().body("User update activity successfully!");
+
     }
 
 }
